+++ src/Grphp/Client/Strategy/Envoy/StrategyFactory.php
@@ -15,8 +15,38 @@
  * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-$_ENV['TEST_MODE'] = 1;
-error_reporting(E_ALL | E_STRICT);
-date_default_timezone_set('UTC');
-require dirname(__DIR__) . '/tests/Support/Client.php';
-require dirname(__DIR__) . '/tests/Support/TestInterceptors.php';
+declare(strict_types = 1);
+
+namespace Grphp\Client\Strategy\Envoy;
+
+use Grphp\Protobuf\Serializer;
+
+/**
+ * Builds Envoy strategies and its dependency graph from the Envoy config
+ */
+class StrategyFactory
+{
+    /** @var Config */
+    private $config;
+
+    /**
+     * @param Config $config
+     */
+    public function __construct(Config $config)
+    {
+        $this->config = $config;
+    }
+
+    /**
+     * @return Strategy
+     */
+    public function build(): Strategy
+    {
+        $serializer = new Serializer();
+        return new Strategy(
+            $serializer,
+            new RequestFactory($this->config, $serializer),
+            new RequestExecutor()
+        );
+    }
+}